+++ src/app/component/particles-js/index.js
@@ -1,0 +1,101 @@
+import { useRecoilValue } from "recoil";
+import { backgroundColor, constellation } from "../../stateManagement/recoil";
+import { useCallback } from 'react'
+import { loadSlim } from "tsparticles-slim";
+import Particles from "react-particles";
+
+export const Particlesjs = () => {
+
+    const bgColor = useRecoilValue(backgroundColor);
+    const asterisms = useRecoilValue(constellation);
+
+    const particlesInit = useCallback(async engine => {
+        console.log(engine);
+        // you can initiate the tsParticles instance (engine) here, adding custom shapes or presets
+        // this loads the tsparticles package bundle, it's the easiest method for getting everything ready
+        // starting from v2 you can add only the features you need reducing the bundle size
+        //await loadFull(engine);
+        await loadSlim(engine);
+    }, []);
+    const particlesLoaded = useCallback(async container => {
+        await console.log(container);
+    }, []);
+
+    return (
+        <Particles
+            id="tsparticles"
+            init={particlesInit}
+            loaded={particlesLoaded}
+            options={{
+                fullScreen: true,
+                background: {
+                    color: {
+                        value: bgColor,
+                    },
+                },
+                fpsLimit: 180,
+                interactivity: {
+                    events: {
+                        onClick: {
+                            enable: true,
+                            mode: "push",
+                        },
+                        onHover: {
+                            enable: true,
+                            mode: "repulse",
+                        },
+                        resize: true,
+                    },
+                    modes: {
+                        push: {
+                            quantity: 4,
+                        },
+                        repulse: {
+                            distance: 200,
+                            duration: 0.4,
+                        },
+                    },
+                },
+                particles: {
+                    color: {
+                        value: asterisms,
+                    },
+                    links: {
+                        color: asterisms,
+                        distance: 150,
+                        enable: true,
+                        opacity: 0.5,
+                        width: 1,
+                    },
+                    move: {
+                        direction: "none",
+                        enable: true,
+                        outModes: {
+                            default: "bounce",
+                        },
+                        random: false,
+                        speed: 6,
+                        straight: false,
+                    },
+                    number: {
+                        density: {
+                            enable: true,
+                            area: 800,
+                        },
+                        value: 80,
+                    },
+                    opacity: {
+                        value: 0.5,
+                    },
+                    shape: {
+                        type: "circle",
+                    },
+                    size: {
+                        value: { min: 3, max: 6 },
+                    },
+                },
+                detectRetina: true,
+            }}
+        />
+    )
+}
