--- src/app/particles-js/index.js
@@ -1,69 +1,0 @@
-export const options = {
-    background: {
-        color: {
-            value: "#ffff",
-        },
-    },
-    fpsLimit: 180,
-    interactivity: {
-        events: {
-            onClick: {
-                enable: true,
-                mode: "push",
-            },
-            onHover: {
-                enable: true,
-                mode: "repulse",
-            },
-            resize: true,
-        },
-        modes: {
-            push: {
-                quantity: 4,
-            },
-            repulse: {
-                distance: 200,
-                duration: 0.4,
-            },
-        },
-    },
-    particles: {
-        color: {
-            value: "#0000",
-        },
-        links: {
-            color: "#000",
-            distance: 150,
-            enable: true,
-            opacity: 0.5,
-            width: 1,
-        },
-        move: {
-            direction: "none",
-            enable: true,
-            outModes: {
-                default: "bounce",
-            },
-            random: false,
-            speed: 6,
-            straight: false,
-        },
-        number: {
-            density: {
-                enable: true,
-                area: 800,
-            },
-            value: 80,
-        },
-        opacity: {
-            value: 0.5,
-        },
-        shape: {
-            type: "circle",
-        },
-        size: {
-            value: { min: 3, max: 6 },
-        },
-    },
-    detectRetina: true,
-}
